+++ src/day01/part01.hpp
@@ -2,7 +2,7 @@
 #include <vector>
 #include <string>
 namespace day01 {
-    long long solve_part1(const std::vector<std::string>& lines);
+    long solve_part1(const std::vector<std::string>& lines);
     int shift_position(const char& side, int& position, const int& shift);
 }
